+++ packages/components/ContextualMenu/src/Submenu.tsx
@@ -1,10 +1,10 @@
 /** @jsx withSlots */
 import * as React from 'react';
-import { View, ScrollView, Platform } from 'react-native';
+import { View, ScrollView, Platform, I18nManager } from 'react-native';
 import { submenuName, SubmenuProps, SubmenuSlotProps, SubmenuType, SubmenuRenderData, SubmenuState } from './Submenu.types';
 import { settings } from './Submenu.settings';
 import { IUseComposeStyling, compose } from '@uifabricshared/foundation-compose';
-import { IFocusable, useSelectedKey } from '@fluentui-react-native/interactive-hooks';
+import { IFocusable, useKeyDownProps, useSelectedKey } from '@fluentui-react-native/interactive-hooks';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
 import { backgroundColorTokens, borderTokens } from '@fluentui-react-native/tokens';
 import { Callout } from '@fluentui-react-native/callout';
@@ -43,11 +43,13 @@ export const Submenu = compose<SubmenuType>({
       userProps?.onShow && userProps.onShow();
       context.isSubmenuOpen = true;
     }, [context]);
+
     const onDismiss = React.useCallback(() => {
       userProps?.onDismiss();
       setShowMenu(false);
       context.isSubmenuOpen = false;
     }, [context, setShowMenu]);
+
     const dismissCallback = React.useCallback(() => {
       onDismiss();
       context?.onDismissMenu();
@@ -70,6 +72,19 @@ export const Submenu = compose<SubmenuType>({
 
     const styleProps = useStyling(userProps, (override: string) => state[override] || userProps[override]);
 
+    const dismissWithArrowKey = React.useCallback(
+      (e: any) => {
+        const arrowKey = I18nManager.isRTL ? 'ArrowRight' : 'ArrowLeft';
+        if (e.nativeEvent.key === arrowKey) {
+          onDismiss();
+        }
+      },
+      [onDismiss],
+    );
+
+    // Explicitly override onKeyDown to override the native windows behavior of moving focus with arrow keys.
+    const onKeyDownProps = useKeyDownProps(dismissWithArrowKey, 'ArrowLeft', 'ArrowRight');
+
     const containerPropsWin32: IViewProps = {
       accessible: shouldFocusOnContainer,
       focusable: shouldFocusOnContainer && containerFocus,
@@ -84,10 +99,10 @@ export const Submenu = compose<SubmenuType>({
         onDismiss: onDismiss,
         setInitialFocus: shouldFocusOnMount,
       },
-      container: Platform.select({
-        macos: {},
-        default: containerPropsWin32,
-      }),
+      container: {
+        ...onKeyDownProps,
+        ...(Platform.OS === ('win32' as string) && { containerPropsWin32 }),
+      },
       scrollView: {
         contentContainerStyle: {
           flexDirection: 'column',