+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,30 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Mon, 07 Feb 2022 19:47:54 GMT and should not be manually modified.
+This log was last generated on Thu, 10 Feb 2022 17:13:16 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.1
+
+Thu, 10 Feb 2022 17:13:16 GMT
+
+### Patches
+
+- Flip arrow key navigation in RTL (sanajmi@microsoft.com)
+- Don't dismiss submenu on menu item click (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/callout to v0.19.24
+
+## 0.16.0
+
+Wed, 09 Feb 2022 20:14:50 GMT
+
+### Minor changes
+
+- Arrow Right/Left open and close the submenu. (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/callout to v0.19.23
+- Bump @fluentui-react-native/focus-zone to v0.9.1
+- Bump @fluentui-react-native/interactive-hooks to v0.13.0
+- Bump @fluentui-react-native/pressable to v0.8.19
 
 ## 0.15.2